requirements.txt
langchain
langchain-chroma
langchain-huggingface
langchain-openai
langgraph
pydantic
python-dotenv
chromadb
